3.26(survey api기능 추가전 업데이트)

+++ src/main/java/com/ezen/propick/user/repository/UserRepository.java
@@ -3,7 +3,4 @@ package com.ezen.propick.user.repository;
 import com.ezen.propick.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
-
+public interface UserRepository extends JpaRepository<User, Integer> {}
